+++ src/Hit/Core.hs
@@ -3,6 +3,7 @@
 
 module Hit.Core
        ( PushBool (..)
+       , CommitOptions (..)
        ) where
 
 
@@ -11,3 +12,17 @@ data PushBool
     = Simple
     | Force
     deriving (Show, Eq)
+
+-- | Options of the @hit commit@ command.
+data CommitOptions = CommitOptions
+    { {- | Commit name. If not specified use the issue name.
+        If issue number is not applicable do not perform any actions.
+        -}
+      coName          :: !(Maybe Text)
+      -- | Do not use the issue num in the commit name.
+    , coNoIssueNumber :: !Bool
+      -- | Push immediately.
+    , coPush          :: !Bool
+      -- | Use Force push?
+    , coIsForcePush   :: !PushBool
+    }